backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
psycopg2-binary
sqlalchemy